+++ include/ignition/rendering/base/BaseMaterial.hh
@@ -271,10 +271,13 @@ namespace ignition
       // \sa Material::SetFragmentShader(const std::string &)
       public: virtual void SetFragmentShader(const std::string &_path) override;
 
+      // Documentation inherited.
       public: virtual void CopyFrom(ConstMaterialPtr _material) override;
 
+      // Documentation inherited.
       public: virtual void CopyFrom(const common::Material &_material) override;
 
+      // Documentation inherited.
       public: virtual void PreRender() override;
 
       protected: virtual void Reset();
@@ -317,6 +320,9 @@ namespace ignition
 
       /// \brief Set to true to enable object with this material to cast shadows
       protected: bool castShadows = true;
+
+      /// \brief Default pbr material properties
+      private: const common::Pbr kDefaultPbr;
     };
 
     //////////////////////////////////////////////////
@@ -859,6 +865,17 @@ namespace ignition
       this->ClearNormalMap();
       // TODO(anyone): update common::Material
       this->SetShaderType(ST_PIXEL);
+
+      const common::Pbr *pbrMat = _material.PbrMaterial();
+      if (!pbrMat)
+        pbrMat = &this->kDefaultPbr;
+      this->SetNormalMap(pbrMat->NormalMap());
+      this->SetRoughnessMap(pbrMat->RoughnessMap());
+      this->SetMetalnessMap(pbrMat->MetalnessMap());
+      this->SetRoughness(pbrMat->Roughness());
+      this->SetMetalness(pbrMat->Metalness());
+      this->SetEnvironmentMap(pbrMat->EnvironmentMap());
+      this->SetEmissiveMap(pbrMat->EmissiveMap());
     }
 
     //////////////////////////////////////////////////
@@ -898,8 +915,8 @@ namespace ignition
       this->ClearRoughnessMap();
       this->ClearMetalnessMap();
       this->ClearEmissiveMap();
-      this->SetRoughness(1.0);
-      this->SetMetalness(0.8);
+      this->SetRoughness(this->kDefaultPbr.Roughness());
+      this->SetMetalness(this->kDefaultPbr.Metalness());
       this->SetShaderType(ST_PIXEL);
     }
     }